pyproject.toml
[project]
name = "system-wizyjny"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "numpy==2.2.6",
    "opencv-python==4.12.0.88",
    "python-snap7==2.0.2",
]
